Structure. move making absolute path to utils.fs from module/install.py

+++ utils/fs.py
@@ -2,16 +2,20 @@ import os
 
 
 def create_symlink(file, symlink_file, force = False):
-    file = os.path.abspath(file)
-    symlink_file = os.path.abspath(os.path.expanduser(symlink_file))
 
+    # check values
+    if not os.path.isabs(file):
+        raise ValueError(file," isn't absolute path")
+    if not os.path.isabs(symlink_file):
+        raise ValueError(symlink_file," isn't absolute path")
     if not os.path.isfile(file):
-        raise ValueError('{} is not a file'.format(file))
+        raise ValueError(file, " is not a file")
 
+    # make nested dirs
     destination_dir = os.path.dirname(symlink_file)
     os.makedirs(destination_dir, mode = 0o755, exist_ok = True)
 
-
+    # remove symlink/file(force) if exist
     if os.path.isfile(symlink_file):
         if force or os.path.islink(symlink_file):
             os.remove(symlink_file)
@@ -19,6 +23,26 @@ def create_symlink(file, symlink_file, force = False):
             print("file {} is exist".format(symlink_file))
             return 0
 
+    # create symlink
     os.symlink(file, symlink_file)
     return 1
 
+def create_list_of_symlink(symlink_pairs, force = False):
+
+    ret_val = True
+    for config, symlink in symlink_pairs:
+        if create_symlink(config, symlink, force):
+            print("{}: symlink was succesfully created".format(config))
+        else:
+            print("{}: symlink wasn't created".format(config))
+            ret_val = False
+
+    return ret_val
+
+def make_absolute_path(pairs, config_prefix, symlink_prefix):
+    config_prefix = os.path.abspath(config_prefix)
+    symlink_prefix = os.path.abspath(symlink_prefix)
+
+    for pair in pairs:
+        pair[0] = os.path.join(config_prefix, pair[0])
+        pair[1] = os.path.join(symlink_prefix, pair[1])
